-Fixed navbar ui issues; -Renamed all search page attributes from "dashboard" to "search" for consistency; -Fixed app crash when navigating to the search fragment.
