+++ app/src/main/res/layout/fragment_search.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.dashboard.DashboardFragment"
+    tools:context=".ui.search.SearchFragment"
     android:background="@color/light_orange"
     android:orientation="vertical">
