+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -3,18 +3,18 @@
 
     <item
         android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
+        android:icon="@drawable/home"
         android:title="@string/title_home" />
 
     <item
         android:id="@+id/navigation_dashboard"
         android:icon="@drawable/find"
-        android:title="@string/title_dashboard"
+        android:title="@string/title_search"
         />
 
     <item
         android:id="@+id/navigation_profile"
-        android:icon="@drawable/ic_notifications_black_24dp"
+        android:icon="@drawable/profile"
         android:title="@string/title_profile" />
 
 </menu>
