+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">ChooseYourOwnBeer</string>
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
+    <string name="title_search">Search</string>
     <string name="title_profile">Profile</string>
     <string name="our_beers">Our beers</string>
     <string name="go_to_detail">Go to detail</string>
